new power: Ultimate dataset class splits data in real time, much faster
Old dataset class used to load data from preprocessed numpy arrays, but
somehow, it got very slow. Maybe the deserialization process was too
expensive. But now, we only load the CSV files and split it on the go.

We achieved the load of 7000 samples per second in a for loop with new
dataset class. Enough for us.